pyproject.toml
[tool.poetry]
name = "mastermind"
version = "0.1.0"
description = "Implementation of mastermind game through a REST API"
authors = ["Adrian Carayol Orenes <adriancarayol@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "3.11.0"
psycopg2-binary = "^2.8.4"
gunicorn = "^20.0.4"
black = "^19.10b0"
pytest-django = "^3.8.0"
loguru = "^0.4.1"
django-rest-auth = "^0.9.5"
pytest-cov = "^2.8.1"
django = "~2.2"
pylint = "^2.4.4"
mypy = "^0.761"
pylint-django = "^2.0.13"
drf-yasg = "^1.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
